Fix user that show in suggestions widget. Fix bug in view.
Suggestions:
Renamed to Some Users. Now, the user login can not show yourself in this widget.

Follow Button:
When when follow a user, is returned the number of following in json.
Update the value of following on the home page.

Notifications:
when click on the user name go to his profile